feature/45-configurationservice: rename configurationservice to ConfigServer
