+++ qiusuo-root/qiusuo-communityservice/build.gradle
@@ -12,10 +12,10 @@ dependencies {
     annotationProcessor "com.querydsl:querydsl-apt:${queryDslVersion}:jpa"
     annotationProcessor "org.springframework.boot:spring-boot-starter-data-jpa"
 
-
     // Lombok
     compileOnly "org.projectlombok:lombok:${lombokVersion}"
     annotationProcessor "org.projectlombok:lombok:${lombokVersion}"
+
     implementation "org.projectlombok:lombok:${lombokVersion}"
 
     implementation "org.springframework.boot:spring-boot-starter-web"
@@ -27,6 +27,7 @@ dependencies {
     implementation 'com.graphql-java-kickstart:graphiql-spring-boot-starter:7.0.1'
     implementation "javax.inject:javax.inject:1"
     implementation  "javax.xml.bind:jaxb-api:2.3.1"
+    implementation 'org.springframework.cloud:spring-cloud-config-client'
 
     runtime files("config")
