+++ qiusuo-root/qiusuo-communityservice/docker/Dockerfile
@@ -1,0 +1,13 @@
+FROM adoptopenjdk/openjdk11:ubi
+RUN  apk update && apk upgrade && apk add netcat-openbsd && apk add curl
+RUN mkdir -p /usr/local/communityservice
+RUN echo $JAVA_HOME
+RUN cd /tmp/ && \
+    curl -k -LO "http://download.oracle.com/otn-pub/java/jce/8/jce_policy-8.zip" -H 'Cookie: oraclelicense=accept-securebackup-cookie' && \
+    unzip jce_policy-8.zip && \
+    rm jce_policy-8.zip && \
+    yes |cp -v /tmp/UnlimitedJCEPolicyJDK8/*.jar /usr/lib/jvm/java-1.8-openjdk/jre/lib/security/
+ADD  @project.build.finalName@.jar /usr/local/communityservice/
+ADD run.sh run.sh
+RUN chmod +x run.sh
+CMD ./run.sh
