+++ qiusuo-root/qiusuo-configsvr/.settings/org.eclipse.core.resources.prefs
