+++ qiusuo-root/qiusuo-configsvr/src/main/java/com/qiusuo/configsvr/QiuSuoConfigServer.java
@@ -1,4 +1,4 @@
-package com.qiusuo.configurationservice;
+package com.qiusuo.configsvr;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,9 +6,9 @@ import org.springframework.cloud.config.server.EnableConfigServer;
 
 @SpringBootApplication
 @EnableConfigServer
-public class QiuSuoConfigurationService {
+public class QiuSuoConfigServer {
     public static void main(String[] args) {
-        SpringApplication.run(QiuSuoConfigurationService.class, args);
+        SpringApplication.run(QiuSuoConfigServer.class, args);
     }
 }
