+++ qiusuo-root/qiusuo-configsvr/src/main/resources/config/qiusuo-communityservice/application.yml
