+++ qiusuo-root/qiusuo-configsvr/src/main/resources/config/qiusuo-messagingservice/application-dev.yml
